+++ src/Common/Helper.tsx
@@ -10,7 +10,8 @@ import YAML from 'yaml'
 import { ERROR_EMPTY_SCREEN, SortingOrder, TOKEN_COOKIE_NAME, EXCLUDED_FALSY_VALUES, DISCORD_LINK } from './Constants'
 import { ServerErrors } from './ServerError'
 import { toastAccessDenied } from './ToastBody'
-import { AsyncOptions, AsyncState, UseSearchString, scrollableInterface } from './Types'
+import { AsyncOptions, AsyncState, UseSearchString } from './Types'
+import { scrollableInterface } from '../Shared'
 import { DATE_TIME_FORMAT_STRING } from '../Shared'
 import { ReactComponent as ArrowDown } from '../Assets/Icon/ic-chevron-down.svg'
