+++ src/Common/Types.ts
@@ -926,6 +926,3 @@ export interface DeploymentWindowProfileMetaData {
     warningMessage: string
 }
 
-export interface scrollableInterface {
-    autoBottomScroll: boolean
-}
