+++ src/Shared/Components/CICDHistory/cicdHistory.scss
@@ -118,3 +118,91 @@
         }
     }
 }
+
+.historical-diff {
+    display: grid;
+    grid-template-rows: 60px auto;
+    height: 100%;
+    background-color: var(--white);
+}
+
+.historical-diff__container {
+    display: grid;
+    grid-template-columns: 280px auto;
+    height: calc(100vh - 147px);
+}
+
+.deployment-diff__upper {
+    display: grid;
+    grid-template-columns: 50% 50%;
+    grid-template-rows: auto;
+}
+
+.historical-diff__left {
+    padding: 16px 12px;
+    overflow: hidden;
+    min-width: 280px;
+}
+
+.history_diff__wrapper {
+    .tab-list__tab-link.active {
+        color: var(--N900);
+        background: #e5f2ff;
+        font-weight: 600;
+        border-bottom: 0px;
+    }
+}
+
+.code-editor-header-value {
+    border-bottom: 1px solid var(--N200);
+    border-top-right-radius: 4px;
+    border-top-left-radius: 4px;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+}
+
+.compare-history__border-right {
+    border-right: 1px solid var(--N100);
+}
+
+.compare-history__border-left {
+    border-left: 1px solid var(--N100);
+}
+
+.configuration-link {
+    &.active {
+        background: var(--B100);
+        color: var(--N900);
+        font-weight: 600;
+        border-right: 1px solid var(--N200);
+    }
+}
+.historical-name:hover {
+    color: var(--N900) !important;
+}
+
+.code-editor-green-diff {
+    background: #eaf1dd;
+}
+
+.code-editor-red-diff {
+    background: #ffd4d1;
+}
+
+.left-50 {
+    left: 50px;
+}
+
+.deployment-aborted {
+    height: calc(100vh - 250px);
+    justify-content: center;
+}
+
+.icon-selected {
+    background-color: #0066cc;
+    border-radius: 4px;
+    path {
+        stroke: #fff;
+    }
+}
