+++ src/Shared/Components/ShowMoreText/ShowMoreText.tsx
